+++ packages/core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @nodeboot/test
 
+## 2.12.0
+
+### Minor Changes
+
+-   Add Test hooks for MongoDB in memory Database and ReplicaSet
+
 ## 2.11.0
 
 ### Minor Changes